third_party/WebKit/LayoutTests/platform/win7/virtual/outofblink-cors-http/http/tests/fetch/serviceworker-proxied/thorough/auth-nocors-other-https-expected.txt

This is a testharness.js-based test.
PASS Startup 
FAIL Verify access control of fetch() in a Service Worker assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?Auth jsonpResult must be success expected "success" but got "error"
Harness: the test ran to completion.
